[RDRViews] need to fix rdr name problem.
