+++ src/semantic_world/views/__init__.py
@@ -1,0 +1,2 @@
+from .view_classifiers import *
+from .views import *
